microblog/requirements.txt
Django==2.0
django-autoslug==1.9.4
getmac==0.8.1
isort==4.3.21
Pillow==6.1.0
pytz==2019.1
